Move file format tests into a new test module
